+++ examples/tsne/responsive_tsne/responsive_tsne.cpp
@@ -37,8 +37,8 @@ float getEEFactor(int iter, const Config& config) {
     return 1.0f;
 }
 // Perform Responsive t-SNE with Progressive KDTree
-void ResponsiveTSNE::run(double* X, int N, int D, double* Y, int no_dims, double perplexity, double theta, int rand_seed,
-    bool skip_random_init, int max_iter, int stop_lying_iter, int mom_switch_iter, Config& config) {
+void ResponsiveTSNE::run(double* X, size_t N, size_t D, double* Y, size_t no_dims, double perplexity, double theta, int rand_seed,
+    bool skip_random_init, size_t max_iter, size_t stop_lying_iter, size_t mom_switch_iter, Config& config) {
 
     // Set random seed
     if (skip_random_init != true) {
